LayoutTests/crypto/subtle/X25519-generate-key-exctractable-expected.txt

Test generating an extractable EC key pair with Curve25519 using X25519 algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key pair...
PASS keyPair.toString() is '[object Object]'
PASS keyPair.publicKey.type is 'public'
PASS keyPair.publicKey.extractable is true
PASS keyPair.publicKey.algorithm.name is 'X25519'
PASS keyPair.publicKey.usages is [ ]
PASS keyPair.privateKey.type is 'private'
PASS keyPair.privateKey.extractable is true
PASS keyPair.privateKey.algorithm.name is 'X25519'
PASS keyPair.privateKey.usages is ['deriveBits', 'deriveKey']
PASS successfullyParsed is true

TEST COMPLETE
